I think someof my commits were never made, so I will combine them here.
I added random punctuation and paragraph generation. It can be controlled by how many words you want to consider a sentence or a paragraph. I added more code to strip punctuation and numbers from the original code, because it was interfering with how I was picking the pairs for the trigrams.